test: refactor fns used in testing
Move to new module, forcing fns to rely
only on public API of ForwardGraph.
